A100/godatadriven/neurips-llm-efficiency-challenge/pyproject.toml
[tool.poetry]
name = "cajajejo"
version = "0.1.0"
description = "Source code for Neurips competition 2023 <https://llm-efficiency-challenge.github.io/>"
authors = [
    "Caio Bernatti <caio.bernatti@xebia.com>",
    "Jordi Smit <jordi.smit@xebia.com>",
    "Jetze Schuurmans <jetze.schuurmans@xebia.com>",
    "Yke Rusticus <yke.rusticus@xebia.com>",
    "Jasper Ginn <jasper.ginn@xebia.com>"
]
readme = "README.md"
packages = [
  { include = "cajajejo", from = "src" }
]

[tool.poetry.scripts]
cajajejo = "cajajejo.cli:entrypoint"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
typer = "^0.9.0"
ruamel-yaml = "~0.17.32"
pyyaml = "~6.0.1"
pydantic = "<2"
jinja2 = "~3.1.2"
sh = "~2.0.4"
coolname = "~2.2.0"
google-cloud-secret-manager = "^2.16.2"
pandas = "^2.1.1"
# NB: needs to go here else 'extras' don't work
mlflow = {version="2.5.0", optional=true}
google-cloud-storage = {version = "~2.10.0", optional = true}
trl = {version = "^0.5.0", optional = true}
transformers = {git = "https://github.com/huggingface/transformers.git", rev = "68e85fc822097b3df8d685a4705804348245284d", optional = true}
accelerate = {git = "https://github.com/huggingface/accelerate.git", rev = "658492fb410396d8d1c1241c1cc2412a908b431b", optional = true}
datasets = {version = "^2.14.4", optional = true}
bitsandbytes = {version = "^0.41.1", optional = true}
einops = {version = "^0.6.1", optional = true}
peft = {git = "https://github.com/huggingface/peft.git", rev = "52ff0cde9f2cc64059e171c2cfd94512914c85df", optional = true}
pyarrow = {version = "<=10.0.0", optional = true}
# NB: required by MLFlow
pytorch-triton = {url = "https://download.pytorch.org/whl/nightly/pytorch_triton-2.1.0%2B6e4932cda8-cp39-cp39-linux_x86_64.whl", optional = true, platform = "linux"}
torch = {url = "https://download.pytorch.org/whl/nightly/cu118/torch-2.2.0.dev20230912%2Bcu118-cp39-cp39-linux_x86_64.whl", optional = true, platform = "linux"}
torchvision = {url = "https://download.pytorch.org/whl/nightly/cu118/torchvision-0.17.0.dev20230912%2Bcu118-cp39-cp39-linux_x86_64.whl", optional = true, platform = "linux"}
sentencepiece = "^0.1.99"
fastapi = {version = "^0.103.1", optional = true}
uvicorn = {extras = ["default"], version = "^0.23.2", optional = true}

huggingface-hub = {version = "^0.17.3", optional = true}
protobuf = ">=3.20.2,<3.21.0"

[tool.poetry.group.dev.dependencies]
invoke = "^2.2.0"
black = "^23.7.0"
ruff = "^0.0.278"
mypy = "^1.4.1"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
python-dotenv = "^1.0.0"
pre-commit = "^3.3.3"
ipykernel = "^6.24.0"
gitpython = "^3.1.35"
pendulum = "^2.1.2"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocstrings = {extras = ["python"], version = "^0.20.0"}
mkdocs-include-markdown-plugin = "^4.0.4"
mkdocs-jupyter = "^0.22.0"
mkdocs-material = "<9.0.0"

[tool.poetry.group.docs]
optional = true

# For pip install only, e.g. "pip install .[evaluation]"
[tool.poetry.extras]
evaluation = ["mlflow", "google-cloud-storage", "transformers"]
training = [
  "mlflow", "google-cloud-storage", "trl", "transformers",
  "accelerate", "datasets", "bitsandbytes", "einops", "peft",
  "pyarrow", "torch", "pandas", "huggingface-hub"
]
submission = ["uvicorn", "fastapi"]

[[tool.poetry.source]]
name = "pytorch-nightly"
url = "https://download.pytorch.org/whl/nightly/cu118"
priority = "explicit"

[tool.black]
line-length = 88
exclude = '''
^/(
  (
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.venv
    | _build
    | build
    | dist
    | .notebooks
    | .nox
  )
)
'''

[tool.ruff]
ignore = ["E501"]
extend-exclude = [
  "__pycache__",
  "docs",
  "site",
  ".nox",
  ".notebooks",
  "toy-submission"
]

[tool.mypy]
exclude = [
  "^.notebooks/",
  "^tasks\\.py$",
]
ignore_missing_imports = true

[tool.poetry-git-version-plugin]
alpha_version_format = '{version}+{commit_hash}'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
